user_management/requirements.txt
aiohttp
flask==2.3.3
Flask-Migrate==2.7.0
redis==5.0.4
rq==1.16.2
Flask-Assets
Flask-Compress
Flask-Login
Flask-Mail
Flask-WTF
raygun4py
WTForms-SQLAlchemy
email-validator
redis-py-cluster
Flask-SQLAlchemy==3.0.5 
pytz
gunicorn
gevent
flask_rq2
flask_socketio
phonenumbers
PyJWT==2.8.0
playwright
cryptography
Flask-CORS
Flask-SSE
psycopg2-binary
yapf
isort
pyotp
qrcode[pil]
qrcodegen
SQLAlchemy==1.4.46